make exped UI response when the fleet is empty
this patch ensures the expedition checker tab works properly
even if the whole fleet (including flagship) is dismissed.

the code is writen with an early version of KanColleHelpers in mind.
that version assumes every fleet should at least have a flagship,
which is not true. despite this problem being fixed long ago in lib,
the code in KC3Kai is never updated to handle the case when fleet is
empty.