provide option for removing wrong remodel info
Some users' remodelDb (the one that keeps track of remodel-related info)
would say remodel level for I13 and I14 are both Lv. 35, which is wrong
(in fact it's Lv.45). This is because KC devs included a placeholder
data for both I13 and I14, and that wronged data is recorded by remodelDb
and kept until next time the length of mst_ship or mst_gear
changes, i.e. something new gets implemented or removed.

there's no other way than issuing some console commands to reset that
data, so I create this link under profile tab which does just that.