Fix Akashi repair prediction
The old logic for predicting the current state of a repair - i.e. amount
of HP that can be restored and how long until the next point can be
repaired - was not rounding correctly, causing incorrect results in some
circumstances.

Additionally, the business logic has been split from the display logic
and moved to AkashiRepair.js