Fix Akashi Repair tooltip when unknown startTime
Previously, if the start time of an Akashi repair was unknown, the
tooltip would be stuck at 0 HP repaired and 20 minutes to next repair.
This could occur if the devtools panel was reinitialized for some
reason, e.g. updating KC3.

Now it will properly indicate that the repaired HP and time are unknown.