option for increased font sizes in Moonlight
They're not really desirable to have increased at high resolutions/sizes, so it was set as an option.

Renamed the bestgirl thing to "partner" within the file structure for no good reason. I have no idea how to implement it properly, so whatever at that. Will probably just disable it completely.