Update ShowcaseExporter.js for the changed icon system
Completely overlooked this file.

Image quality is a problem with the new icons if no smoothing method is used, but if someone has a better method or implementation, you know what to do. Probably need to find a proper method of converting into circles.

Also, ongoing HD icon adjustments for the circle format. (~65%)