Fix ledger not working with very large history
The kc3 ledger suddenly stops working if your resource history is extensive enough. This is because current behavior of using apply with newTotalBuffer calls unshift with as many args as there are items in your resource history. The javascript engine does not like this when that number starts reaching hundreds of thousands.

To fix, we slice newTotalBuffer and prepend it in batches.