moonlight v2 prototype + syncup with natsuiro
This is a big, giant, clunky mess. (understatement) The entire thing was probably a mistake, in so many ways.

I was working on trying to support skins in moonlight back at the start of 2020, but then I had to go on a hiatus. I have since resumed the project.

It is nowhere near completion.

`settings` names are temporary. Some settings are likely to be removed, and there are other planned settings that aren't implemented.

There is a reasonable chance that the whole thing will have to be scrapped.